New batch of fixes
