+++ lib/nebula_web/views/v1/domain_view.ex
@@ -2,7 +2,7 @@ defmodule NebulaWeb.V1.DomainView do
   use NebulaWeb, :view
   require Logger
 
-  def render("cdmi_domain.json", %{cdmi_domain: cdmi_domain}) do
-    cdmi_domain
+  def render("cdmi_domain.json", %{object: object}) do
+    object
   end
 end
